Re-enabling the rest of the tests and adapting them to the new design Further refactoring is still needed, lots of duplicate code

+++ src/sudoku/view/event/FinishedSolvingEvent.java
@@ -7,7 +7,7 @@ import sudoku.model.SudokuModel;
  */
 public class FinishedSolvingEvent extends SudokuEvent {
 
-    public FinishedSolvingEvent(SudokuModel sudoku, long time) {
+    public FinishedSolvingEvent(SudokuModel sudoku, float time) {
         super(SudokuEventType.FinishedSolving, sudoku);
         this.message = "Solved " + sudoku.getName() + " in " + (time / 1000) + "s";
         this.time = time;
